glorious funtoo updates
merged:

  bar-overlay: 1ad50aaef1e93863ab64d2b36a4a28686d55bb17
  bliss-overlay: 51003cf9be3629eb4805769828adb3cb10e2d209
  squeezebox: cf7953988018d60d3638245b380406be494bd7ae
  funtoo-deadbeef: 387e40259abfabc35512865e807d055b75f8bf87
  funtoo-redhat: 3f60fd94264c2ad1e5909b5c1825e51854e81cad
  funtoo-wmfs: 9f97ae675241901f470402ac27373b11512982ea
  vmware: b98ffa6d5d97a41a8b7be0a258a5adf92ed2de27
  miramir: a3265040c5cc4c3acdc5815c60debbe71d587a42
  sera: 97b0bfd1f11402ba2185f151efa5fbd1641601fa
  faustoo: 031fa6de6bd560b2d39c1f36a13114715296be4b
  foo-overlay: 44af5419fabdb2f70469f6696330eac1c37b60d9
  funtoo-plex: 19e6c68cebc0f438e6784b91dfd0ad70b71647dd
  causes: 9b263e0ba528cd8e66dc1b13462de4de115e6fab
  sabayon-for-gentoo: 74de53e379c50c17c20814c52d23cf7d53d08c5b
  funtoo-gnome: a4b6d7ddb29e96b3f9d3e2c2dae22d164cabb140
  funtoo-mysql: d5d7016fc384bfd95091c0372ab8d8a011fd411c
  funtoo-media: 4cec88ede97c64486543ac127245480a401597be
  funtoo-overlay: 04ff0a240b502678479552611842e226ef7beb3b
  funtoo-toolchain: 81df4551c3d57be041e3ab498af12256857b88c1
  funtoo-ldap: b97b3fd7ab9c0ef7ab7884a57c0d9f1609b6b999

+++ sys-apps/systemd/systemd-217-r4.ebuild
@@ -1,6 +1,6 @@
 # Copyright 1999-2014 Gentoo Foundation
 # Distributed under the terms of the GNU General Public License v2
-# $Header: /var/cvsroot/gentoo-x86/sys-apps/systemd/systemd-217-r1.ebuild,v 1.1 2014/11/04 19:24:27 floppym Exp $
+# $Header: /var/cvsroot/gentoo-x86/sys-apps/systemd/systemd-217-r4.ebuild,v 1.2 2014/12/19 17:36:41 pacho Exp $
 
 EAPI=5
 
@@ -12,14 +12,15 @@ inherit autotools-utils bash-completion-r1 linux-info multilib \
 
 DESCRIPTION="System and service manager for Linux"
 HOMEPAGE="http://www.freedesktop.org/wiki/Software/systemd"
-SRC_URI="http://www.freedesktop.org/software/systemd/${P}.tar.xz"
+SRC_URI="http://www.freedesktop.org/software/systemd/${P}.tar.xz
+	http://dev.gentoo.org/~floppym/dist/systemd-gentoo-patches-217-r1.tar.xz"
 
 LICENSE="GPL-2 LGPL-2.1 MIT public-domain"
 SLOT="0/2"
 KEYWORDS="~alpha ~amd64 ~arm ~ia64 ~ppc ~ppc64 ~sparc ~x86"
 IUSE="acl apparmor audit cryptsetup curl doc elfutils gcrypt gudev http
 	idn introspection kdbus +kmod lz4 lzma pam policykit python qrcode +seccomp
-	selinux ssl terminal test vanilla"
+	selinux ssl sysv-utils terminal test vanilla"
 
 MINKV="3.8"
 
@@ -47,6 +48,9 @@ COMMON_DEPEND=">=sys-apps/util-linux-2.25:0=
 	qrcode? ( media-gfx/qrencode:0= )
 	seccomp? ( sys-libs/libseccomp:0= )
 	selinux? ( sys-libs/libselinux:0= )
+	sysv-utils? (
+		!sys-apps/systemd-sysv-utils
+		!sys-apps/sysvinit )
 	terminal? ( dev-libs/libevdev:0=
 		>=x11-libs/libxkbcommon-0.4:0=
 		x11-libs/libdrm:0= )
@@ -58,6 +62,7 @@ RDEPEND="${COMMON_DEPEND}
 	>=sys-apps/baselayout-2.2
 	!sys-auth/nss-myhostname
 	!<sys-libs/glibc-2.14
+	!sys-fs/eudev
 	!sys-fs/udev"
 
 # sys-apps/dbus: the daemon only (+ build-time lib dep for tests)
@@ -90,6 +95,7 @@ src_prepare() {
 	cp "${FILESDIR}"/217-systemd-consoled.service.in \
 		units/user/systemd-consoled.service.in || die
 
+	EPATCH_FORCE=yes EPATCH_SUFFIX=patch epatch
 	autotools-utils_src_prepare
 }
 
@@ -318,10 +324,16 @@ multilib_src_install_all() {
 	prune_libtool_files --modules
 	einstalldocs
 
-	# we just keep sysvinit tools, so no need for the mans
-	rm "${D}"/usr/share/man/man8/{halt,poweroff,reboot,runlevel,shutdown,telinit}.8 \
-		|| die
-	rm "${D}"/usr/share/man/man1/init.1 || die
+	if use sysv-utils; then
+		for app in halt poweroff reboot runlevel shutdown telinit; do
+			dosym "..${ROOTPREFIX}/bin/systemctl" /sbin/${app}
+		done
+	else
+		# we just keep sysvinit tools, so no need for the mans
+		rm "${D}"/usr/share/man/man8/{halt,poweroff,reboot,runlevel,shutdown,telinit}.8 \
+			|| die
+		rm "${D}"/usr/share/man/man1/init.1 || die
+	fi
 
 	# Disable storing coredumps in journald, bug #433457
 	mv "${D}"/usr/lib/sysctl.d/50-coredump.conf{,.disabled} || die
